+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 3.0.0-nullsafety
+
+* Migrate package to null-safe dart.
+* Most parameters of the `FakePlatform` constructor are now required to match
+  non-nullability of `Platform` APIs.
+
 ### 2.2.1
 
 * Add `operatingSystemVersion`